hackflow/backend/pyproject.toml
[tool.poetry]
name = "hackflow-backend"
version = "0.1.0"
description = "HackFlow — Hackathon Automation Platform"
authors = ["HackFlow Team"]
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.0"
redis = {extras = ["asyncio"], version = "^5.0.4"}
aiosmtplib = "^3.0.1"
jinja2 = "^3.1.4"
chromadb = "^0.5.0"
google-generativeai = "^0.7.0"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
